+++ src/main/java/com/meoguri/linkocean/controller/bookmark/CategoryController.java
@@ -5,7 +5,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.meoguri.linkocean.domain.bookmark.entity.vo.Category;
-import com.meoguri.linkocean.support.controller.ListResponse;
+import com.meoguri.linkocean.support.controller.dto.ListResponse;
 
 @RequestMapping("/api/v1/categories")
 @RestController
