+++ src/main/java/com/meoguri/linkocean/controller/notification/NotificationController.java
@@ -10,7 +10,7 @@ import org.springframework.web.bind.annotation.RestController;
 import com.meoguri.linkocean.configuration.security.jwt.SecurityUser;
 import com.meoguri.linkocean.domain.notification.entity.Notification;
 import com.meoguri.linkocean.domain.notification.service.NotificationService;
-import com.meoguri.linkocean.support.controller.SliceResponse;
+import com.meoguri.linkocean.support.controller.dto.SliceResponse;
 
 import lombok.RequiredArgsConstructor;
