+++ src/main/java/com/meoguri/linkocean/support/controller/dto/ListResponseJsonSerializer.java
@@ -1,4 +1,4 @@
-package com.meoguri.linkocean.support.controller;
+package com.meoguri.linkocean.support.controller.dto;
 
 import java.io.IOException;
 
@@ -7,26 +7,23 @@ import com.fasterxml.jackson.databind.JsonSerializer;
 import com.fasterxml.jackson.databind.SerializerProvider;
 
 /**
- * 페이지 조회의 공통 Serializer
+ * 리스트 조회의 공통 Serializer
  * - 데이터 타입 별로 array 의 이름을 커스터마이징 하기 위해 사용
  */
-public class PageResponseJsonSerializer extends JsonSerializer<PageResponse<?>> {
+public class ListResponseJsonSerializer extends JsonSerializer<ListResponse<?>> {
 
 	@Override
-	public void serialize(final PageResponse value, final JsonGenerator gen, final SerializerProvider serializers)
+	public void serialize(final ListResponse<?> value, final JsonGenerator gen, final SerializerProvider serializers)
 		throws IOException {
 
 		gen.writeStartObject();
-		gen.writeFieldName("totalCount");
-		gen.writeNumber(value.getTotalCount());
-		gen.writeArrayFieldStart(value.getName());
 
+		gen.writeArrayFieldStart(value.getName());
 		for (Object datum : value.getData()) {
 			gen.writeObject(datum);
 		}
-
 		gen.writeEndArray();
+
 		gen.writeEndObject();
 	}
-
 }
