+++ src/main/java/com/meoguri/linkocean/support/controller/dto/PageResponseJsonSerializer.java
@@ -1,4 +1,4 @@
-package com.meoguri.linkocean.support.controller;
+package com.meoguri.linkocean.support.controller.dto;
 
 import java.io.IOException;
 
@@ -7,26 +7,25 @@ import com.fasterxml.jackson.databind.JsonSerializer;
 import com.fasterxml.jackson.databind.SerializerProvider;
 
 /**
- * 슬라이스 조회의 공통 Serializer
+ * 페이지 조회의 공통 Serializer
  * - 데이터 타입 별로 array 의 이름을 커스터마이징 하기 위해 사용
  */
-public class SliceResponseJsonSerializer extends JsonSerializer<SliceResponse<?>> {
+public class PageResponseJsonSerializer extends JsonSerializer<PageResponse<?>> {
 
 	@Override
-	public void serialize(final SliceResponse value, final JsonGenerator gen, final SerializerProvider serializers)
+	public void serialize(final PageResponse value, final JsonGenerator gen, final SerializerProvider serializers)
 		throws IOException {
 
 		gen.writeStartObject();
-
-		gen.writeFieldName("hasNext");
-		gen.writeBoolean(value.getHasNext());
-
+		gen.writeFieldName("totalCount");
+		gen.writeNumber(value.getTotalCount());
 		gen.writeArrayFieldStart(value.getName());
+
 		for (Object datum : value.getData()) {
 			gen.writeObject(datum);
 		}
-		gen.writeEndArray();
 
+		gen.writeEndArray();
 		gen.writeEndObject();
 	}
